pyproject.toml
[tool.poetry]
name = "telegram-user-tracker"
version = "0.1.0"
description = ""
authors = ["Hung-I Wang <whygowe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
telethon = "^1.16.4"
cryptg = "^0.2.post1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
rope = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
